backend/requirements.txt
beautifulsoup4
requests
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain
langchain-core
langchain-openai
langchain-chroma
langchain-community
chromadb
streamlit
selenium
webdriver-manager
PyJWT 
python-multipart
jwt
dotenv
pydantic[email]
psycopg2-binary